refactor: 캐러셀 문구 변경 및 캐러셀 꽉 채우도록 코드 수정
