+++ src/pages/user/Main.tsx
@@ -5,44 +5,9 @@ import useProductAllQuery from '@/hooks/reactQuery/product/useProductAllQuery';
 import useFilterByCategory from '@/hooks/category/useFilterByCategory';
 import useDisplayCount from '@/hooks/functionHooks/useDispalyControl';
 import { useUserStore } from '@/utils/Zustand';
-import carousel1 from '@/assets/images/carousel1.jpg';
-import carousel2 from '@/assets/images/carousel2.jpg';
-import carousel3 from '@/assets/images/carousel3.jpg';
-import Carousel from '@/components/main/Carousel';
 import Layout from '@/components/common/layout/Layout';
 import MainCard from '@/components/main/MainCard';
 import PartnerMain from '@/pages/partner/PartnerMain';
-
-interface ImageItem {
-  url: string;
-  text?: string;
-  text2?: string;
-  path?: string;
-  click?: string;
-}
-
-const carousel: ImageItem[] = [
-  {
-    url: carousel1,
-    text: '머물고 싶은, 즐기고싶은',
-    text2: '장소를 큐레이팅합니다',
-    path: '/',
-  },
-  {
-    url: carousel2,
-    text: '프라이빗 풀이 있는',
-    text2: '추천 숙소 9곳',
-    path: '/list/1',
-    click: '숙소 상품 둘러보기',
-  },
-  {
-    url: carousel3,
-    text: '자연에서 즐기는',
-    text2: '카파도키아 열기구 투어',
-    path: '/list/2',
-    click: '체험 상품 둘러보기',
-  },
-];
 
 const Main = () => {
   const navigate = useNavigate();
@@ -97,8 +62,7 @@ const Main = () => {
   }
 
   return (
-    <Layout main category noSearch={false}>
-      <Carousel items={carousel} />
+    <Layout main category noSearch={false} carouselView>
       <div className="flex justify-center p-20 pb-5 mb-80">
         <MainCard
           images={CategoryAccommodation}
